fix sample compressed tiling
