+++ hub/core/chunk/sample_compressed_chunk.py
@@ -7,6 +7,7 @@ from hub.core.serialize import (
     bytes_to_text,
     check_sample_size,
 )
+from hub.core.tiling.sample_tiles import SampleTiles
 from .base_chunk import BaseChunk
 
 SampleValue = Union[Sample, np.ndarray, int, float, bool, dict, list, str]
@@ -25,19 +26,28 @@ class SampleCompressedChunk(BaseChunk):
                 incoming_sample, self.compression, self.is_byte_compression
             )
             self.num_dims = self.num_dims or len(shape)
-            sample_nbytes = len(serialized_sample)
             check_sample_shape(shape, self.num_dims)
-            check_sample_size(sample_nbytes, self.min_chunk_size, self.compression)
-            if serialized_sample:
-                # optimization so that even if this sample doesn't fit, it isn't recompressed next time we try
-                incoming_samples[i] = Sample(
-                    buffer=serialized_sample, compression=self.compression, shape=shape
-                )
-            if not self.can_fit_sample(sample_nbytes):
+            if isinstance(serialized_sample, SampleTiles):
+                incoming_samples[i] = serialized_sample
+                if not self.data_bytes:
+                    self.write_tile(serialized_sample)
+                    num_samples += 0.5
                 break
-            self.data_bytes += serialized_sample
-            self.register_in_meta_and_headers(sample_nbytes, shape)
-            num_samples += 1
+            else:
+                sample_nbytes = len(serialized_sample)
+                check_sample_size(sample_nbytes, self.min_chunk_size, self.compression)
+                if serialized_sample:
+                    # optimization so that even if this sample doesn't fit, it isn't recompressed next time we try
+                    incoming_samples[i] = Sample(
+                        buffer=serialized_sample,
+                        compression=self.compression,
+                        shape=shape,
+                    )
+                if not self.can_fit_sample(sample_nbytes):
+                    break
+                self.data_bytes += serialized_sample
+                self.register_in_meta_and_headers(sample_nbytes, shape)
+                num_samples += 1
         return num_samples
 
     def read_sample(
