+++ hub/core/tiling/sample_tiles.py
@@ -1,4 +1,5 @@
 import numpy as np
+from hub.core.compression import compress_array
 from hub.core.tiling.optimizer import get_tile_shape
 from hub.core.tiling.serialize_tile import break_into_tiles, serialize_tiles
 
@@ -15,7 +16,9 @@ class SampleTiles:
             arr.shape, arr.nbytes * compression_ratios[compression], chunk_size, -1
         )
         tiles = break_into_tiles(arr, self.tile_shape)
-        self.tiles = serialize_tiles(tiles, lambda x: memoryview(x.tobytes()))
+        self.tiles = serialize_tiles(
+            tiles, lambda x: memoryview(compress_array(x, self.compression))
+        )
         self.registered = False
         # product of self.tiles.shape
         self.num_tiles = self.tiles.size
